use (unsigned?) longs for lengths

+++ trunk/src/audio/datasource_interface.h
@@ -12,7 +12,7 @@ class IDataSource {
 		virtual bool exhausted() = 0;
 		virtual void reset() = 0;
 
-		virtual int read(char* const buffer, int len) = 0;
+		virtual int read(char* const buffer, unsigned long len) = 0;
 };
 
 #endif//DATASOURCE_INTERFACE_H
